+++ .env.example
@@ -1,3 +1,4 @@
 TELEGRAM_TOKEN='12345'
 SECRET_KEY='secret_key_django'
 BASE_DIR=/путь/до/директории/где/manage.py
+ADMIN_ID=123
