+++ README.md
@@ -16,6 +16,9 @@ https://github.com/vityn101979
 
 <h2>Как использовать</h2>
 
+Заполните файл .env как в .env.example.
+(насчет admin_id не знаю нужен ли он(нужны ли функции с ним), но с ним все рабоает)
+
 запуск бота:
 
 `cd core`
