+++ core/culture/management/commands/bot.py
@@ -6,7 +6,7 @@ from aiogram import Bot, Dispatcher
 from aiogram.exceptions import TelegramServerError
 
 from .config import TELEGRAM_TOKEN, logger
-from .commands import form_router
+from .commands import main_router
 
 
 async def main():
@@ -15,7 +15,7 @@ async def main():
     except TelegramServerError as error:
         logger.error(f'Ошибка в запуске бота {error}')
     dp = Dispatcher()
-    dp.include_router(form_router)
+    dp.include_router(main_router)
     # не уверен что надо иммено тут
     logger.info("Bot включился!")
     await dp.start_polling(bot, allowed_updates=dp.resolve_used_update_types())
