+++ core/culture/management/commands/handlers/__init__.py
@@ -1,0 +1,2 @@
+from .meetings import meetings_router # noqa
+from .route import route_router # noqa
